[C] Use ‘in_port_t’ instead of ‘unsigned short’.
U ≡ "Convert type of arg or local var
either from ‘unsigned short *’ to ‘in_port_t *’,
or from ‘unsigned short’ to ‘in_port_t’";
S(X) ≡ "Likewise, for struct member(s) X";
C ≡ "Remove cast to ‘unsigned short’".
* src/libserveez/binding.c (svz_binding_filter_net): U.
(svz_sock_local_info, svz_binding_filter): U.
* src/libserveez/core.c (svz_socket_connect): U.
* src/libserveez/core.h (svz_socket_connect): U.
* src/libserveez/icmp-socket.c (svz_icmp_header_t): S (port).
(svz_icmp_connect): U; C.
* src/libserveez/icmp-socket.h (svz_icmp_connect): U.
* src/libserveez/portcfg.c (svz_portcfg_port): U.
* src/libserveez/portcfg.h
(svz_portcfg_t): S (u.tcp.port, u.udp.port).
(svz_portcfg_port): U.
* src/libserveez/socket.c (svz_sock_intern_connection_info): U.
* src/libserveez/socket.h (svz_socket_t): S (remote_port, local_port).
* src/libserveez/tcp-socket.c (svz_tcp_connect): U.
* src/libserveez/tcp-socket.h (svz_tcp_connect): U.
* src/libserveez/udp-socket.c (svz_udp_connect): U.
* src/libserveez/udp-socket.h (svz_udp_connect): U.

U ≡ "Convert type of arg or local var
either from ‘unsigned short *’ to ‘in_port_t *’,
or from ‘unsigned short’ to ‘in_port_t’";
S(X) ≡ "Likewise, for struct member(s) X";
C ≡ "Remove cast to ‘unsigned short’".
* src/awcs-server/awcs-proto.c (awcs_status_connected): U.
* src/guile-api.c (guile_sock_connect): U; C.
(guile_svz_ntohs, guile_svz_htons, guile_sock_remote_address)
(guile_sock_local_address, scm_portmap): C.
* src/guile.c (guile_define_port): C.
* src/irc-server/irc-proto.h (irc_server_t): S (port).
* src/irc-server/irc-server.c (irc_connect_servers): C.
* src/gnutella.c (nut_connect_ip): U.
(dns_closure): S (port).
(nut_dns_done, nut_connect_host): U.
(nut_init): C.
* src/nut-server/gnutella.h (nut_pong_t)
(nut_reply_t, nut_push_t, nut_host_t, nut_config_t): S (port).
* src/nut-server/nut-core.c (nut_parse_host, nut_parse_addr): U; C.
(nut_client_key): U.
* src/nut-server/nut-core.h (nut_parse_host)
(nut_parse_addr, nut_client_key): U.
* src/nut-server/nut-hostlist.c (nut_host_catcher): U.
* src/nut-server/nut-hostlist.h (nut_host_catcher): U.
* src/nut-server/nut-request.c (nut_query, nut_ping): C.
* src/nut-server/nut-transfer.c (nut_send_push): C.
* src/tunnel-server/tunnel.c (tnl_create_socket): U.
* src/tunnel-server/tunnel.h (tnl_connect_t): S (port).

+++ src/guile.c
@@ -1178,7 +1178,7 @@ guile_define_port (SCM name, SCM args)
       err |= optionhash_extract_int (options, PORTCFG_PORT, 0, 0,
                                      &port, action);
       GUILE_VALIDATE_PORT (port, "TCP", portname);
-      SVZ_CFG_TCP (cfg, port) = (unsigned short) port;
+      SVZ_CFG_TCP (cfg, port) = port;
       err |= optionhash_extract_int (options, PORTCFG_BACKLOG, 1, 0,
                                      &SVZ_CFG_TCP (cfg, backlog), action);
       err |= optionhash_extract_string (options, PORTCFG_IP, 1,
@@ -1195,7 +1195,7 @@ guile_define_port (SCM name, SCM args)
       err |= optionhash_extract_int (options, PORTCFG_PORT,
                                      0, 0, &port, action);
       GUILE_VALIDATE_PORT (port, "UDP", portname);
-      SVZ_CFG_UDP (cfg, port) = (unsigned short) port;
+      SVZ_CFG_UDP (cfg, port) = port;
       err |= optionhash_extract_string (options, PORTCFG_IP, 1,
                                         SVZ_PORTCFG_NOIP,
                                         &SVZ_CFG_UDP (cfg, ipaddr), action);
